Create G.F Series.py
Siddhant made a special series and named it as G.F Series. The series follows this trend  Tn=(Tn-2)2-(Tn-1)  in which the first and the second term are 0 and 1 respectively. Help Siddhant to find the first N terms of the series.

Note: Print a white space (" ") after every integer and a newline character after every testcase. The generated output is white space sensitive, do not add extra spaces on unnecessary newline characters.